Tests moved, and doc framework finished
