--- src/docs/doc-components/ComponentDocs/ComponentDocs.tsx
@@ -1,11 +1,0 @@
-import * as React from "react";
-
-interface IComponentDocs {
-  componentName: string;
-}
-
-export class ComponentDocs extends React.Component<IComponentDocs> {
-  public render() {
-    return <div>{this.props.componentName}</div>;
-  }
-}
