--- src/docs/doc-components/ComponentDocs/index.tsx
@@ -1,1 +1,0 @@
-export * from "./ComponentDocs";
