+++ src/docs/doc-components/HelperComponents/_index.sass
@@ -1,0 +1,62 @@
+@import "../variables"
+
+.atlas-sidebar
+    display: flex
+    flex-direction: column
+    flex: 1 0 20%
+    background: $white
+    margin-top: 10px
+    border-radius: 5px
+    border: 1px solid $faded-dark-gray
+
+    .header
+        display: flex
+        justify-content: center
+        padding: 10px 0
+        border-bottom: 1px solid $faded-dark-gray
+
+    .component-list
+        display: flex
+        flex: 1
+        flex-direction: column
+        height: 100%
+
+        .component-row
+            display: flex
+            align-items: center
+            height: 20px
+            font-size: 15px
+            cursor: pointer
+            padding: 10px
+            padding-left: 20px
+            transition: all 250ms ease-in-out
+
+        .component-row:hover
+            color: $blue
+
+        .selected
+            background-color: $golden
+            height: 30px
+
+.component-holder
+    display: flex
+    flex: 1 0 80%
+    margin: 10px 10px 0 10px
+    padding: 10px
+    background: $white
+    border-radius: 10px
+    border: 1px solid $faded-dark-gray
+
+    .main-title
+        font-size: 25px
+        animation: fade-in 1.5s 1
+        
+
+@keyframes fade-in
+    0%
+        opacity: 0
+    100%
+        opacity: 1
+
+
+
