+++ src/docs/doc-components/HelperComponents/index.tsx
@@ -1,0 +1,2 @@
+export * from "./ComponentHolder";
+export * from "./Navbar";
