--- src/docs/doc-components/Navbar/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Navbar";
